added access to mesh bounding box data
