+++ tests/data/vm-11111111-1111-1111-1111-111111111111/meshBoundaryBoundingBox.json
@@ -1,0 +1,98 @@
+{
+    "blk-1/BODY": {
+        "xmax": 3.804954767227173,
+        "xmin": -0.8078516721725464,
+        "ymax": 1.0149916410446167,
+        "ymin": -1.0149916410446167,
+        "zmax": 1.1024686098098755,
+        "zmin": -0.17886872589588165
+    },
+    "blk-1/WT_ceiling": {
+        "xmax": 41.80443572998047,
+        "xmin": -14.810705184936523,
+        "ymax": 6.891486167907715,
+        "ymin": -6.891366481781006,
+        "zmax": 9.69450855255127,
+        "zmin": 9.694380760192871
+    },
+    "blk-1/WT_ground": {
+        "xmax": 41.80443572998047,
+        "xmin": -3.1070165634155273,
+        "ymax": 6.891486167907715,
+        "ymin": -6.89132833480835,
+        "zmax": -0.30551815032958984,
+        "zmin": -0.30561956763267517
+    },
+    "blk-1/WT_ground_close": {
+        "xmax": 5.892999172210693,
+        "xmin": -3.1070005893707275,
+        "ymax": 2.0,
+        "ymin": -2.0,
+        "zmax": -0.30551815032958984,
+        "zmin": -0.30553847551345825
+    },
+    "blk-1/WT_ground_front": {
+        "xmax": -3.1069846153259277,
+        "xmin": -14.810705184936523,
+        "ymax": 6.891339302062988,
+        "ymin": -6.891366481781006,
+        "zmax": -0.30549171566963196,
+        "zmin": -0.30551815032958984
+    },
+    "blk-1/WT_ground_patch": {
+        "xmax": 2.9879958629608154,
+        "xmin": -0.16114698350429535,
+        "ymax": 0.9065545797348022,
+        "ymin": -0.9065545797348022,
+        "zmax": -0.29636433720588684,
+        "zmin": -0.30553191900253296
+    },
+    "blk-1/WT_inlet": {
+        "xmax": -14.810705184936523,
+        "xmin": -14.810705184936523,
+        "ymax": 6.891301155090332,
+        "ymin": -6.891366481781006,
+        "zmax": 9.69450855255127,
+        "zmin": -0.30549171566963196
+    },
+    "blk-1/WT_outlet": {
+        "xmax": 41.80443572998047,
+        "xmin": 41.80436325073242,
+        "ymax": 6.891486167907715,
+        "ymin": -6.891181945800781,
+        "zmax": 9.694380760192871,
+        "zmin": -0.30561956763267517
+    },
+    "blk-1/WT_side1": {
+        "xmax": 41.80443572998047,
+        "xmin": -14.810705184936523,
+        "ymax": -6.891181945800781,
+        "ymin": -6.891366481781006,
+        "zmax": 9.69450855255127,
+        "zmin": -0.30561956763267517
+    },
+    "blk-1/WT_side2": {
+        "xmax": 41.80436325073242,
+        "xmin": -14.810705184936523,
+        "ymax": 6.891486167907715,
+        "ymin": 6.891301155090332,
+        "zmax": 9.69450855255127,
+        "zmin": -0.30561956763267517
+    },
+    "blk-1/wheel_rim": {
+        "xmax": 3.0145559310913086,
+        "xmin": -0.21436680853366852,
+        "ymax": 0.8657052516937256,
+        "ymin": -0.8657052516937256,
+        "zmax": 0.22136951982975006,
+        "zmin": -0.22137364745140076
+    },
+    "blk-1/wheel_tire": {
+        "xmax": 3.111682176589966,
+        "xmin": -0.3114983141422272,
+        "ymax": 0.876271665096283,
+        "ymin": -0.876271665096283,
+        "zmax": 0.3184996545314789,
+        "zmin": -0.30351021885871887
+    }
+}
