+++ tests/data/vm-1cfdec99-3ce3-428c-85f8-2054812b2ddc/meshBoundaryBoundingBox.json
@@ -1,0 +1,1 @@
+{"rotatingBlock-cylinder_blade1/blade1": {"xmax": 6.710000000000008, "xmin": 1.0199199999999995, "ymax": 0.17134244202569274, "ymin": -0.5140716519483178, "zmax": 0.03953287490211811, "zmin": -0.05499992790351302}, "rotatingBlock-cylinder_blade1/slidingInterface-cylinder_blade1": {"xmax": 7.25, "xmin": 0.25, "ymax": 0.899913046235743, "ymin": -0.899913046235743, "zmax": 0.8996522017450986, "zmin": -0.9}, "rotatingBlock-cylinder_blade2/blade2": {"xmax": -1.0199199999999995, "xmin": -6.710000000000006, "ymax": 0.5140716519483186, "ymin": -0.1713424420256919, "zmax": 0.039532874902102985, "zmin": -0.05499992790351302}, "rotatingBlock-cylinder_blade2/slidingInterface-cylinder_blade2": {"xmax": -0.25, "xmin": -7.25, "ymax": 0.899913046235743, "ymin": -0.899913046235743, "zmax": 0.9, "zmin": -0.8996522017450986}, "rotatingBlock-main_rotor/slidingInterface-cylinder_blade1": {"xmax": 7.25, "xmin": 0.25, "ymax": 0.899913046235743, "ymin": -0.899913046235743, "zmax": 0.8996522017450986, "zmin": -0.9}, "rotatingBlock-main_rotor/slidingInterface-cylinder_blade2": {"xmax": -0.25, "xmin": -7.25, "ymax": 0.899913046235743, "ymin": -0.899913046235743, "zmax": 0.9, "zmin": -0.8996522017450986}, "rotatingBlock-main_rotor/slidingInterface-main_rotor": {"xmax": 8.0, "xmin": -7.999373376577781, "ymax": 7.999843342610599, "ymin": -7.999843342610599, "zmax": 1.5, "zmin": -1.5000000000000007}, "stationaryBlock/farfield": {"xmax": 670.3536154860747, "xmin": -670.3601907422825, "ymax": 670.4804965137522, "ymin": -670.2058403051981, "zmax": 670.3645622698933, "zmin": -670.7070896167738}, "stationaryBlock/slidingInterface-main_rotor": {"xmax": 8.0, "xmin": -7.999373376577781, "ymax": 7.999843342610599, "ymin": -7.999843342610599, "zmax": 1.5, "zmin": -1.5000000000000007}}
